requirements.txt
aniso8601==9.0.1
click==8.1.3
dnspython==2.3.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
opencv-python==4.7.0.68
PyJWT==2.6.0
pymongo==4.3.3
pyserial==3.5
pytz==2022.7.1
six==1.16.0
Werkzeug==2.2.2
zipp==3.12.0